refactor: useThemeContext 생성
