+++ FE/src/context/themeContext.jsx
@@ -1,10 +1,22 @@
-import React, { createContext, useState } from 'react';
+import React, { createContext, useContext, useState } from 'react';
 import PropTypes from 'prop-types';
+
 import { ThemeProvider as StyledProvider } from 'styled-components';
+
 import { lightColors, darkColors } from '../styles/color';
 import { fontSize, fontWeight } from '../styles/font';
 
 const ThemeContext = createContext({});
+
+const useThemeContext = () => {
+  const theme = useContext(ThemeContext);
+
+  if (Object.keys(theme) === 0) {
+    throw new Error('useThemeContext should be used within ThemeProvider');
+  }
+
+  return theme;
+};
 
 const ThemeProvider = ({ children }) => {
   const localTheme = window.localStorage.getItem('theme') || 'light';
@@ -22,6 +34,4 @@ ThemeProvider.propTypes = {
   children: PropTypes.any,
 };
 
-export { ThemeContext, ThemeProvider };
-
-export default ThemeProvider;
+export { useThemeContext, ThemeProvider };
